+++ CMakeLists.txt
@@ -9,7 +9,13 @@ set(CMAKE_CXX_COMPILER clang++)
 add_library(reig_lib
         lib/reig.h
         lib/reig.cpp
-        lib/stb_truetype.h)
+        lib/reig/stb_truetype.h
+        lib/reig/fwd.h
+        lib/reig/defs.h
+        lib/reig/primitive.h
+        lib/reig/mouse.h
+        lib/reig/exception.h
+        lib/reig/reference_widget.h lib/reig/context.h)
 add_library(my::reig_lib ALIAS reig_lib)
 target_compile_features(reig_lib
         PUBLIC
