+++ lib/reig.cpp
@@ -3,7 +3,7 @@
 #define STB_TRUETYPE_IMPLEMENTATION
 #define STBTT_STATIC
 
-#include "stb_truetype.h"
+#include "reig/stb_truetype.h"
 #include <sstream>
 #include <memory>
