+++ lib/reig.h
@@ -1,581 +1,11 @@
 #ifndef REIG_H_INCLUDED
 #define REIG_H_INCLUDED
 
-#include "stb_truetype.h"
-#include <vector>
-#include <cstdint>
-#include <any>
-#include <string>
-
-namespace reig {
-    using uint8_t = std::uint8_t;
-    using uint32_t = std::uint32_t;
-
-    namespace primitive {
-        struct Point {
-            float x = 0.0f;
-            float y = 0.0f;
-        };
-
-        struct Rectangle {
-            float x = 0.0f;
-            float y = 0.0f;
-            float width = 0.0f;
-            float height = 0.0f;
-        };
-
-        struct Triangle {
-            Point pos0;
-            Point pos1;
-            Point pos2;
-        };
-
-        struct Color {
-            struct Red {
-                uint8_t val = 0u;
-
-                constexpr explicit Red(uint8_t val = 0u) noexcept
-                        : val{val} {}
-            };
-
-            struct Green {
-                uint8_t val = 0u;
-
-                constexpr explicit Green(uint8_t val = 0u) noexcept
-                        : val{val} {}
-            };
-
-            struct Blue {
-                uint8_t val = 0u;
-
-                constexpr explicit Blue(uint8_t val = 0u) noexcept
-                        : val{val} {}
-            };
-
-            struct Alpha {
-                uint8_t val = 0xFFu;
-
-                constexpr explicit Alpha(uint8_t val = 0xFFu) noexcept
-                        : val{val} {}
-            };
-
-            constexpr Color() = default;
-
-            constexpr Color(Red const& red, Green const& green, Blue const& blue,
-                            Alpha const& alpha = Alpha{0xFFu}) noexcept
-                    : red{red}, green{green}, blue{blue}, alpha{alpha} {}
-
-            Red red;
-            Green green;
-            Blue blue;
-            Alpha alpha;
-        };
-
-        namespace colors {
-            uint32_t to_uint(Color const& from);
-
-            Color from_uint(uint32_t rgba);
-
-            namespace literals {
-                constexpr Color::Red operator "" _r(unsigned long long val) noexcept {
-                    return Color::Red{static_cast<uint8_t>(val)};
-                }
-
-                constexpr Color::Green operator "" _g(unsigned long long val) noexcept {
-                    return Color::Green{static_cast<uint8_t>(val)};
-                }
-
-                constexpr Color::Blue operator "" _b(unsigned long long val) noexcept {
-                    return Color::Blue{static_cast<uint8_t>(val)};
-                }
-
-                constexpr Color::Alpha operator "" _a(unsigned long long val) noexcept {
-                    return Color::Alpha{static_cast<uint8_t>(val)};
-                }
-            }
-
-            namespace operators {
-                namespace detail {
-                    template <typename Comp>
-                    static constexpr bool is_color_component_v = std::is_same_v<Comp, Color::Red> ||
-                                                                 std::is_same_v<Comp, Color::Green> ||
-                                                                 std::is_same_v<Comp, Color::Blue> ||
-                                                                 std::is_same_v<Comp, Color::Alpha>;
-
-                    template <typename Comp, typename = std::enable_if_t<is_color_component_v<Comp>>>
-                    Comp const& get_comp(Color const& color) {
-                        if constexpr (std::is_same_v<Comp, Color::Red>) return color.red;
-                        else if constexpr (std::is_same_v<Comp, Color::Green>) return color.green;
-                        else if constexpr (std::is_same_v<Comp, Color::Blue>) return color.blue;
-                        else return color.alpha;
-                    }
-
-                    template <typename Comp, typename = std::enable_if_t<is_color_component_v<Comp>>>
-                    Comp& get_comp(Color& color) {
-                        if constexpr (std::is_same_v<Comp, Color::Red>) return color.red;
-                        else if constexpr (std::is_same_v<Comp, Color::Green>) return color.green;
-                        else if constexpr (std::is_same_v<Comp, Color::Blue>) return color.blue;
-                        else return color.alpha;
-                    }
-                }
-
-                template <typename Comp, typename = std::enable_if_t<detail::is_color_component_v<Comp>>>
-                Color operator|(Color const& left, Comp const& right) {
-                    Color ret = left;
-                    detail::get_comp<Comp>(ret) = right;
-                    return ret;
-                };
-
-                template <typename Comp, typename = std::enable_if_t<detail::is_color_component_v<Comp>>>
-                Color operator+(Color const& left, Comp const& right) {
-                    Color ret = left;
-                    detail::get_comp<Comp>(ret).val += right.val;
-                    return ret;
-                };
-
-                template <typename Comp, typename = std::enable_if_t<detail::is_color_component_v<Comp>>>
-                Color operator-(Color const& left, Comp const& right) {
-                    Color ret = left;
-                    detail::get_comp<Comp>(ret).val -= right.val;
-                    return ret;
-                };
-            }
-
-            // @formatter:off
-            #pragma clang diagnostic push
-            #pragma ide diagnostic ignored "OCUnusedGlobalDeclarationInspection"
-            Color constexpr transparent{Color::Red{},    Color::Green{},    Color::Blue{},    Color::Alpha{}};
-            Color constexpr red        {Color::Red{239}, Color::Green{41},  Color::Blue{41}};
-            Color constexpr orange     {Color::Red{252}, Color::Green{175}, Color::Blue{62}};
-            Color constexpr yellow     {Color::Red{252}, Color::Green{233}, Color::Blue{79}};
-            Color constexpr green      {Color::Red{138}, Color::Green{226}, Color::Blue{52}};
-            Color constexpr blue       {Color::Red{114}, Color::Green{159}, Color::Blue{207}};
-            Color constexpr violet     {Color::Red{173}, Color::Green{127}, Color::Blue{168}};
-            Color constexpr brown      {Color::Red{143}, Color::Green{89},  Color::Blue{2}};
-            Color constexpr white      {Color::Red{255}, Color::Green{255}, Color::Blue{255}};
-            Color constexpr lightGrey  {Color::Red{186}, Color::Green{189}, Color::Blue{182}};
-            Color constexpr mediumGrey {Color::Red{136}, Color::Green{138}, Color::Blue{133}};
-            Color constexpr darkGrey   {Color::Red{46},  Color::Green{52},  Color::Blue{54}};
-            Color constexpr black      {Color::Red{0},   Color::Green{0},   Color::Blue{0}};
-            #pragma clang diagnostic pop
-            // @formatter:on
-        }
-
-        struct Vertex {
-            Point position;
-            Point texCoord;
-            Color color;
-        };
-    }
-
-    /* forward */ class Context;
-
-    namespace primitive {
-        /**
-         * @class Figure
-         * @brief A bunch of vertices and indices to render a figure
-         * Can be collected by the user, but formation is accessible only for the Context
-         */
-        class Figure {
-        public:
-            /**
-             * @brief Returns figure's read-only vertices
-             */
-            std::vector<primitive::Vertex> const& vertices() const;
-
-            /**
-             * @brief Returns figure's read-only indices
-             */
-            std::vector<int> const& indices() const;
-
-            /**
-             * @brief Return figure's texture index
-             */
-            int texture() const;
-
-        private:
-            Figure() = default;
-
-            friend class ::reig::Context;
-
-            void form(std::vector<primitive::Vertex>& vertices, std::vector<int>& indices, int id = 0);
-
-            std::vector<primitive::Vertex> mVertices;
-            std::vector<int> mIndices;
-            int mTextureId = 0;
-        };
-    }
-
-    namespace detail {
-        class MouseButton {
-        public:
-            MouseButton() = default;
-
-            MouseButton(MouseButton const&) = delete;
-
-            MouseButton(MouseButton&&) = delete;
-
-            MouseButton& operator=(MouseButton const&) = delete;
-
-            MouseButton& operator=(MouseButton&&) = delete;
-
-            /**
-             * @brief Sets mouse pressed and clicked states
-             * @param x X coordinate
-             * @param y Y coordinate
-             */
-            void press(float x, float y);
-
-            /**
-             * @brief Unsets mouse pressed state
-             */
-            void release();
-
-            const primitive::Point& get_clicked_pos() const;
-
-            bool is_pressed() const;
-
-            bool is_clicked() const;
-
-        private:
-            friend class ::reig::Context;
-
-            primitive::Point mClickedPos;
-            bool mIsPressed = false;
-            bool mIsClicked = false;
-        };
-
-        class Mouse {
-        public:
-            Mouse() = default;
-
-            Mouse(Mouse const&) = delete;
-
-            Mouse(Mouse&&) = delete;
-
-            Mouse& operator=(Mouse const&) = delete;
-
-            Mouse& operator=(Mouse&&) = delete;
-
-            detail::MouseButton leftButton;
-            detail::MouseButton rightButton;
-
-            /**
-             * @brief Moves cursor against previous position
-             * @param difx Delta x coordinate
-             * @param dify Delta y coordinate
-             */
-            void move(float difx, float dify);
-
-            /**
-             * @brief Places the cursors in abosulute coordinates
-             * @param x X coordinate
-             * @param y Y coordiante
-             */
-            void place(float x, float y);
-
-            /**
-             * @brief Scrolls the virtual mouse wheel
-             * @param dy Amount of scrolling
-             */
-            void scroll(float dy);
-
-            const primitive::Point& get_cursor_pos() const;
-
-            float get_scrolled() const;
-
-        private:
-            friend class ::reig::Context;
-
-            primitive::Point mCursorPos;
-            float mScrolled = 0.0f;
-        };
-    }
-
-    namespace exception {
-        struct FailedToLoadFontException : std::exception {
-        public:
-            const char* what() const noexcept override;
-
-            static FailedToLoadFontException noTextureId(const char* filePath);
-            static FailedToLoadFontException invalidSize(const char* filePath, float fontSize);
-            static FailedToLoadFontException couldNotOpenFile(const char* filePath);
-            static FailedToLoadFontException invalidFile(const char* filePath);
-            static FailedToLoadFontException couldNotFitCharacters(const char* filePath, float fontSize, int width, int height);
-        private:
-            explicit FailedToLoadFontException(std::string message);
-            const std::string message;
-        };
-
-        struct NoRenderHandlerException : std::exception {
-            const char* what() const noexcept override;
-        };
-    }
-
-    namespace reference_widget {
-        struct button {
-            char const* mTitle = "";
-            primitive::Rectangle mBoundingBox;
-            primitive::Color mBaseColor;
-
-            /**
-             * @brief Render a titled button
-             *
-             * @param title Text to be displayed on button
-             * @param box Button's bounding box
-             * @param color Button's base color
-             *
-             * @return True if the button was clicked, false otherwise
-             */
-            bool draw(Context& ctx) const;
-        };
-
-        struct textured_button {
-            char const* mTitle = "";
-            primitive::Rectangle mBoundingBox;
-            int mBaseTexture = 0, mHoverTexture = 0;
-
-            /**
-             * @brief Render a titled textured button
-             * @param box Button's bouding box
-             * @param baseTexture Button's texture index, when idle
-             * @param hoverTexture Button's texture index, when button is hoverred
-             * @return True if the button was clicked, false otherwise
-             */
-            bool draw(Context& ctx) const;
-        };
-
-        struct label {
-            char const* mTitle = "";
-            primitive::Rectangle mBoundingBox;
-            /**
-             * @brief Render a label, which will be enclosed in the current window, if any
-             * @param text Text to be displayed
-             * @param box Text's bounding box
-             */
-            void draw(Context& ctx) const;
-        };
-
-        struct slider {
-            primitive::Rectangle mBoundingBox;
-            primitive::Color mBaseColor;
-            float& mValueRef;
-            float mMin = 0.0f, mMax = 0.0f, mStep = 0.0f;
-
-            /**
-             * @brief Renders a slider.
-             * @param box Slider's bounding box
-             * @param color Slider's base color
-             * @param value A reference to the value to be represented and changed
-             * @param min The lowest represantable value
-             * @param max The highest represantable value
-             * @param step The discrete portion by which the value can change
-             * @return True if value changed
-             */
-            bool draw(Context& ctx) const;
-        };
-
-        struct textured_slider {
-            primitive::Rectangle mBoundingBox;
-            int mBaseTexture = 0, mCursorTexture = 0;
-            float& mValueRef;
-            float mMin = 0.0f, mMax = 0.0f, mStep = 0.0f;
-
-            /**
-             * @brief
-             * @brief Renders a slider.
-             * @param box Slider's bounding box
-             * @param baseTexture Slider's base texture index
-             * @param cursorTexture Slider's cursor texture index
-             * @param value A reference to the value to be represented and changed
-             * @param min The lowest represantable value
-             * @param max The highest represantable value
-             * @param step The discrete portion by which the value can change
-             * @return True if value changed
-             */
-            bool draw(Context& ctx) const;
-        };
-
-        struct scrollbar {
-            primitive::Rectangle mBoundingBox;
-            primitive::Color mBaseColor;
-            float& mValueRef;
-            float mViewSize = 0.0f;
-
-            /**
-             * @brief Renders a vertical scrollbar
-             * @param box Scrollbar's position and size
-             * @param color Checkbox's base color
-             * @param value A reference to the float to be changed
-             * @return True if value changed
-             */
-            bool draw(Context& ctx) const;
-        };
-
-        struct checkbox {
-            primitive::Rectangle mBoundingBox;
-            primitive::Color mBaseColor;
-            bool& mValueRef;
-            /**
-             * @brief Renders a checkbox
-             * @param box Checkbox's position and size
-             * @param color Checkbox's base color
-             * @param value A reference to the bool to be changed
-             * @return True if value changed
-             */
-            bool draw(Context& ctx) const;
-        };
-
-        struct textured_checkbox {
-            primitive::Rectangle mBoundingBox;
-            int mBaseTexture = 0, mCheckTexture = 0;
-            bool& mValueRef;
-
-            /**
-             * @brief Renders a textured checkbox
-             * @param box Checkbox's position and size
-             * @param baseTexture Checkbox's base texture
-             * @param tickTexture Checkbox's tick texture
-             * @param value A reference to the bool to be changed
-             * @return True if value changed
-             */
-            bool draw(Context& ctx) const;
-        };
-    }
-
-    namespace detail {
-        struct Font {
-            std::vector<stbtt_bakedchar> bakedChars;
-            float size = 0.f;
-            int texture = 0;
-            int width = 0;
-            int height = 0;
-        };
-
-        struct Window {
-            std::vector<primitive::Figure> drawData;
-            char const* title = nullptr;
-            float* x = nullptr;
-            float* y = nullptr;
-            float w = 0.f;
-            float h = 0.f;
-            float headerSize = 0.f;
-            bool started = false;
-
-            void expand(primitive::Rectangle& box);
-        };
-    }
-
-    /**
-     * @class Context
-     * @brief Used to pump in input and request gui creation
-     */
-    class Context {
-    public:
-        Context() = default;
-
-        using DrawData = std::vector<primitive::Figure>;
-        using RenderHandler = void (*)(DrawData const&, std::any&);
-
-        /**
-         * @brief Set's a user function, which will draw the gui, based 
-         * @param handler A C function pointer to a rendering callback
-         * The handler should return void and take in DrawData const& and void*
-         */
-        void set_render_handler(RenderHandler handler);
-
-        /**
-         * @brief Set a user pointer to store in the context.
-         * @param ptr A user data pointer.
-         * The pointer is then automatically passed to callbacks.
-         */
-        void set_user_ptr(std::any ptr);
-
-        /**
-         * @brief Gets the stored user pointer
-         */
-        std::any const& get_user_ptr() const;
-
-        struct FontData {
-            std::vector<uint8_t> bitmap;
-            int width = 0;
-            int height = 0;
-        };
-
-        /**
-         * @brief Sets reig's font to be used for labels
-         * @param fontFilePath The path to fonts .ttf file.
-         * @param textureId This id will be passed by reig to render_handler with text vertices
-         * @param fontHeightPx Font's pixel size
-         * @return Returns the bitmap, which is used to create a texture by user.
-         * Set returned bitmap field to nullptr, to avoid deletion
-         */
-        FontData set_font(char const* fontFilePath, int textureId, float fontHeightPx);
-
-        float get_font_height() const;
-
-        /**
-         * @brief Resets draw data and inputs
-         */
-        void start_new_frame();
-
-        /**
-         * @brief Uses stored drawData and draws everything using the user handler
-         */
-        void render_all();
-
-        // Inputs
-        detail::Mouse mouse;
-
-        // Widget renders
-        void start_window(char const* title, float& x, float& y);
-
-        void end_window();
-
-        template <typename T>
-        auto enqueue(T&& t) {
-            return t.draw(*this);
-        }
-
-        void fit_rect_in_window(primitive::Rectangle& rect);
-
-        // Primitive renders
-        /**
-         * @brief Render some text
-         * @param text Text to be displayed
-         * @param box Text's bounding box
-         */
-        void render_text(char const* text, primitive::Rectangle box);
-
-        /**
-         * @brief Schedules a rectangle drawing
-         * @param rect Position and size
-         * @param color Color
-         */
-        void render_rectangle(primitive::Rectangle const& rect, primitive::Color const& color);
-
-        /**
-         * @brief Schedules a textured rectangle drawing (the texture is stretched)
-         * @param rect Position and size
-         * @param texture Index to the texture
-         */
-        void render_rectangle(primitive::Rectangle const& rect, int texture);
-
-        /**
-         * @brief Schedules a triangle drawing
-         * @param triangle Position and size
-         * @param color Color
-         */
-        void render_triangle(primitive::Triangle const& triangle, primitive::Color const& color);
-
-    private:
-        std::vector<primitive::Figure> mDrawData;
-        detail::Font mFont;
-        detail::Window mCurrentWindow;
-
-        RenderHandler mRenderHandler = nullptr;
-        std::any mUserPtr;
-    };
-}
+#include "reig/defs.h"
+#include "reig/primitive.h"
+#include "reig/mouse.h"
+#include "reig/exception.h"
+#include "reig/reference_widget.h"
+#include "reig/context.h"
 
 #endif // REIG_H_INCLUDED
