+++ lib/reig/defs.h
@@ -1,0 +1,11 @@
+#ifndef REIG_INTS_H
+#define REIG_INTS_H
+
+#include <cstdint>
+
+namespace reig {
+    using uint8_t = std::uint8_t;
+    using uint32_t = std::uint32_t;
+}
+
+#endif //REIG_INTS_H
