+++ lib/reig/fwd.h
@@ -1,0 +1,8 @@
+#ifndef REIG_FWD_H
+#define REIG_FWD_H
+
+namespace reig {
+    class Context;
+}
+
+#endif //REIG_FWD_H
